requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
arrow==1.3.0
attrs==25.1.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.4
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
#flask-mysqldb==1.0.1
fpdf==1.7.2
groq==0.28.0
html2pdf==0.1
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
llama_stack==0.1.6
llama_stack_client==0.1.6
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mistune==3.1.2
MouseInfo==0.1.3
mysql-connector-python==9.3.0
numpy==2.2.3
openai==1.90.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pyaml==25.1.0
#PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pygetwindow==0.0.9
Pygments==2.19.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyjokes==0.8.3
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rich==13.9.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.41
svglib==1.5.1
termcolor==2.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uri-template==1.3.0
uritools==5.0.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
xhtml2pdf==0.2.17